broke filter

+++ src/api.js
@@ -2,6 +2,14 @@ import renders from "./renders";
 
 const BASE_URL = 'http://thinkful-list-api.herokuapp.com/steve-henderson/bookmarks';
 
+/* API FUNCTIONS */
+/* 3 Functions
+    getItems();
+    createItem();
+    deleteItem();
+*/
+
+
 const getItems = function () {
   return fetch(`${BASE_URL}`)
 };
@@ -12,17 +20,12 @@ const createItem = function (title, description = " ", url, rating) {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
-      // 'Access-Control-Allow-Origin': "*/*",
-      // "Access-Control-Allow-Credentials": "true",
-      // "Access-Control-Allow-Methods": "GET,HEAD,OPTIONS,POST,PUT",
-      // 'Access-Control-Allow-Headers': "Access-Control-Allow-Origin Access-Control-Allow-Headers, Origin,Accept, X-Requested-With, Content-Type, Access-Control-Request-Method, Access-Control-Request-Headers"
     },
     body: JSON.stringify({
       title: title,
       desc: description,
       url: url,
       rating: Number(rating),
-      //"expanded": false
     })
   })
   .then(res => {
@@ -35,14 +38,22 @@ const createItem = function (title, description = " ", url, rating) {
 };
 
 const deleteItem = function (id) {
-  return fetch(`${BASE_URL}/` + id , {
-    method: 'delete'
-  })
-    .then(response => response.json()
-    .then(json => {
-      return json;
+  if(id === "temp") {
+  getItems()
+      .then(res => res.json())
+      // .then((items) => {
+      //   items.filter()
+    //})
+  } else {
+    return fetch(`${BASE_URL}/` + id , {
+      method: 'delete'
     })
-  );
+      .then(response => response.json()
+      .then(json => {
+        return json;
+      })
+    );
+  };
 };
 
 export default {
